fix: roste_post and employee actioons
