+++ one_fm/one_fm/doctype/post_types_not_filled/post_types_not_filled.json
@@ -28,20 +28,23 @@
   },
   {
    "fieldname": "operations_shift",
-   "fieldtype": "Data",
+   "fieldtype": "Link",
    "ignore_user_permissions": 1,
-   "label": "Operations Shift"
+   "in_list_view": 1,
+   "label": "Operations Shift",
+   "options": "Operations Shift"
   },
   {
    "fieldname": "quantity",
    "fieldtype": "Int",
+   "in_list_view": 1,
    "label": "Quantity"
   }
  ],
  "index_web_pages_for_search": 1,
  "istable": 1,
  "links": [],
- "modified": "2023-03-16 01:20:33.493856",
+ "modified": "2023-05-04 23:20:30.549249",
  "modified_by": "Administrator",
  "module": "one_fm",
  "name": "Post Types Not Filled",
