+++ one_fm/utils.py
@@ -1765,184 +1765,6 @@ def update_onboarding_doc_for_bank_account(doc):
             oe.workflow_state = 'Bank Account'
         oe.save(ignore_permissions=True)
 
-def issue_roster_actions():
-    # Queue roster actions functions to backgrounds jobs
-    frappe.enqueue(create_roster_employee_actions, is_async=True, queue='long')
-    frappe.enqueue(create_roster_post_actions, is_async=True, queue='long')
-
-
-def create_roster_employee_actions():
-    """
-    This function creates a Roster Employee Actions document and issues notifications to relevant supervisors
-    directing them to schedule employees that are unscheduled and assigned to them.
-    It computes employees not scheduled for the span of two weeks, starting from tomorrow.
-    """
-
-    # start date to be from tomorrow
-    start_date = add_to_date(cstr(getdate()), days=1)
-    # end date to be 14 days after start date
-    end_date = add_to_date(start_date, days=14)
-
-    #-------------------- Roster Employee actions ------------------#
-    # fetch employees that are active and don't have a schedule in the specified date range
-    employees_not_rostered = frappe.db.sql("""
-                            select
-                                employee from `tabEmployee`
-                            where
-                                status = 'Active'
-                            AND
-                                employee not in
-                                (select employee
-                                from `tabEmployee Schedule`
-                                where date >= %(start)s and date <=%(end)s)""",
-                                {'start': start_date, 'end': end_date})
-
-    list_of_leaves_within_employee_action_period = frappe.db.get_list("Leave Application", {"status": "Approved", "from_date":["<", start_date ], "to_date": [">", end_date]}, pluck="employee")
-
-    employees = ()
-
-    # fetch employees that are not rostered from the result returned by the query and append to tuple
-    for emp in employees_not_rostered:
-        if emp[0] in list_of_leaves_within_employee_action_period:
-            continue
-        employees = employees + emp
-
-    # fetch supervisors and list of employees(not rostered) under them
-    result = frappe.db.sql("""select sv.employee, sv.site, group_concat(e.employee)
-            from `tabEmployee` e
-            join `tabOperations Shift` sh on sh.name = e.shift
-            join `tabEmployee` sv on sh.supervisor=sv.employee
-            where e.employee in {employees}
-            group by sv.employee """.format(employees=employees))
-
-    # for each supervisor, create a roster action
-    for res in result:
-        supervisor = res[0]
-        site_supervisor = frappe.get_value('Operations Site', res[1], 'account_supervisor')
-        employees = res[2].split(",")
-
-        roster_employee_actions_doc = frappe.new_doc("Roster Employee Actions")
-        roster_employee_actions_doc.start_date = start_date
-        roster_employee_actions_doc.end_date = end_date
-        roster_employee_actions_doc.status = "Pending"
-        roster_employee_actions_doc.action_type = "Roster Employee"
-        roster_employee_actions_doc.supervisor = supervisor
-        roster_employee_actions_doc.site_supervisor = site_supervisor
-
-        for emp in employees:
-            roster_employee_actions_doc.append('employees_not_rostered', {
-                'employee': emp
-            })
-
-        roster_employee_actions_doc.save()
-        frappe.db.commit()
-
-    #-------------------- END Roster Employee actions ------------------#
-
-
-def create_roster_post_actions():
-    """
-    This function creates a Roster Post Actions document that issues actions to supervisors to fill post types that are not filled for a given date range.
-    """
-    
-    op_shift = frappe.db.sql(f""" select supervisor, name from `tabOperations Shift` """)
-    shift_dict = {}
-    for item in op_shift:
-        if item[0] in shift_dict.keys():
-            shift_dict[item[0]].append(item[1])
-        else:
-            shift_dict[item[0]] = [item[1]]
-
-    # start date to be from tomorrow
-    start_date = add_to_date(cstr(getdate()), days=1)
-    # end date to be 14 days after start date
-    end_date = add_to_date(start_date, days=14)
-
-    operations_roles_not_filled_set = set()
-
-    list_of_dict_of_operations_roles_not_filled = []
-    
-    # Fetch post schedules in the date range that are active
-    post_schedules = frappe.db.get_list("Post Schedule", {'date': ['between', (start_date, end_date)], 'post_status': 'Planned'}, ["date", "shift", "operations_role", "post"], order_by="date asc")
-    # Fetch employee schedules for employees who are working
-    employee_schedules = frappe.db.get_list("Employee Schedule", {'date': ['between', (start_date, end_date)], 'employee_availability': 'Working'}, ["date", "shift", "operations_role"], order_by="date asc")
-    
-    for ps in post_schedules:
-        # if there is not any employee schedule that matches the post schedule for the specified date, add to post types not filled
-        if not any(es.date == ps.date and es.shift == ps.shift and es.operations_role == ps.operations_role for es in employee_schedules):
-            if ps.operations_role:
-                # Fetch the project and confirm if the is_active field of the project is set,omit operation roles where the project is not active               
-                project_ = frappe.get_value("Operations Role",ps.operations_role,'project')
-                if project_:
-                    is_active = frappe.get_value("Project",project_,'is_active')
-                    if is_active == "Yes":
-                        operations_roles_not_filled_set.add(ps.operations_role)
-                        list_of_dict_of_operations_roles_not_filled.append(ps)   
-                else:
-                    operations_roles_not_filled_set.add(ps.operations_role)
-                    list_of_dict_of_operations_roles_not_filled.append(ps)
-                    
-
-                
-
-
-    # Convert set to tuple for passing it in the sql query as a parameter
-    operations_roles_not_filled = tuple(operations_roles_not_filled_set)
-    
-    if not operations_roles_not_filled:
-        return
-
-    #Fetch supervisor and post types in his/her shift
-    result = frappe.db.sql("""select sv.employee, group_concat(distinct ps.operations_role)
-            from `tabPost Schedule` ps
-            join `tabOperations Shift` sh on sh.name = ps.shift
-            join `tabEmployee` sv on sh.supervisor=sv.employee
-            where ps.operations_role in {operations_roles}
-            group by sv.employee""".format(operations_roles=operations_roles_not_filled))
-
-
-    # For each supervisor, create post actions to fill post type specifying the post types not filled
-    
-    for res in result:
-        supervisor = res[0]
-        operations_roles = res[1].split(",")
-
-        check_list = []
-        second_check_list = []
-        for val in list_of_dict_of_operations_roles_not_filled:
-            if val["operations_role"] in operations_roles and val["shift"] in shift_dict[supervisor]:
-                check_list.append(val)
-
-        for item in check_list:
-            for second_item in second_check_list:
-                if (item["date"] == second_item["date"]) and (item["shift"] == second_item["shift"]) and (item["operations_role"] == second_item["operations_role"]):
-                    second_item["quantity"] = second_item["quantity"] + 1
-                    break
-            item.update({"quantity": 1})
-            
-            second_check_list.append(item)
-            check_list.remove(item)
-
-        if second_check_list and len(second_check_list) > 0:
-            roster_post_actions_doc = frappe.new_doc("Roster Post Actions")
-            roster_post_actions_doc.start_date = start_date
-            roster_post_actions_doc.end_date = end_date
-            roster_post_actions_doc.status = "Pending"
-            roster_post_actions_doc.action_type = "Fill Post Type"
-            roster_post_actions_doc.supervisor = supervisor
-
-            for obj in second_check_list:
-                roster_post_actions_doc.append('operations_roles_not_filled', {
-                    'operations_role': obj.get("operations_role"),
-                    "operations_shift": obj.get("shift"),
-                    "date": obj.get("date"),
-                    "quantity": obj.get("quantity") if obj.get("quantity") else 1
-                })
-
-            roster_post_actions_doc.save()
-            frappe.db.commit()
-        del check_list
-
 def send_roster_report():
     # Enqueue roster report generation to background
     frappe.enqueue(generate_roster_report, is_async=True, queue='long')
